model 2/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
shapely
python-dotenv
alembic
scikit-learn
openai
httpx
python-multipart
